unvibe this fucking codebase + Tourniquet system + rework the delta packet + redo limb damage + add stimulant + add blood clotters

+++ src/main/java/com/warfactory/medical/mixin/CameraMixin.java
@@ -15,25 +15,14 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 /**
- * CLIENT-ONLY: fixes the THIRD-PERSON (F5) camera on a DOWNED player.
- *
- * <p>The downed first-person camera is placed at the lying head via a low eye-height (0.4) in
- * {@code PlayerMixin#getStandingEyeHeight}. But {@code Camera.setup} uses that SAME eye-height as the pivot the
- * third-person camera orbits and zooms back from — so in third person the orbit point sits at ground level and
- * the zoom-back raytrace ({@code getMaxZoom}) immediately hits the floor, jamming the camera into the ground /
- * the body. This injects right before the detached zoom-back call and raises the pivot to roughly torso height
- * above the player's feet, so the third-person camera orbits the lying body properly. First person (not
- * detached) is untouched and keeps the head-on-ground view.</p>
- *
- * <p>Registered CLIENT-only so {@code Minecraft} / {@code ClientDownedTracker} never load on a dedicated
- * server. The gate is a cheap camera-entity identity + downed-set lookup.</p>
+ * CLIENT-ONLY: fixes the third-person camera on a downed player. The low eye-height (0.4) set by
+ * PlayerMixin also moves the orbit pivot to ground level, which makes the zoom-back raytrace immediately
+ * hit the floor and jam the camera into the body. This injects before the zoom-back and raises the pivot
+ * to torso height; first-person is untouched.
  */
 @Mixin(Camera.class)
 public abstract class CameraMixin {
 
-    /**
-     * Third-person orbit pivot height above the downed player's feet (≈ standing torso height).
-     */
     private static final double WFMEDICAL$DOWNED_TP_PIVOT = 1.0D;
 
     @Shadow
@@ -59,7 +48,7 @@ public abstract class CameraMixin {
         // The camera position is currently the (low, downed) eye pivot; raise it to torso height before the
         // vanilla zoom-back (which runs immediately after this injection point) orbits/zooms from it.
         Vec3 pos = getPosition();
-        double pivotY = Mth.lerp((double) partialTick, entity.yOld, entity.getY()) + WFMEDICAL$DOWNED_TP_PIVOT;
+        double pivotY = Mth.lerp(partialTick, entity.yOld, entity.getY()) + WFMEDICAL$DOWNED_TP_PIVOT;
         setPosition(pos.x, pivotY, pos.z);
     }
 }
